Partial Revert of "Cjs to esm top down (#2313)" excluding dependency cruiser
This reverts commit 578451e1
