+++ backend/__tests__/acceptance/api.info.spec.js
@@ -12,8 +12,8 @@ const { mockRequest } = require('@gardener-dashboard/request')
 describe('api', function () {
   let agent
 
-  beforeAll(async () => {
-    agent = await createAgent()
+  beforeAll(() => {
+    agent = createAgent()
   })
 
   afterAll(() => {
