+++ backend/__tests__/acceptance/api.shoots.spec.js
@@ -22,8 +22,8 @@ function createStore (items) {
 describe('api', function () {
   let agent
 
-  beforeAll(async () => {
-    agent = await createAgent()
+  beforeAll(() => {
+    agent = createAgent()
 
     cache.initialize({
       projects: {
