+++ backend/__tests__/acceptance/api.user.spec.js
@@ -11,8 +11,8 @@ const { mockRequest } = require('@gardener-dashboard/request')
 describe('api', function () {
   let agent
 
-  beforeAll(async () => {
-    agent = await createAgent()
+  beforeAll(() => {
+    agent = createAgent()
   })
 
   afterAll(() => {
