+++ backend/__tests__/acceptance/config.spec.js
@@ -11,8 +11,8 @@ const { mockRequest } = require('@gardener-dashboard/request')
 describe('config', function () {
   let agent
 
-  beforeAll(async () => {
-    agent = await createAgent()
+  beforeAll(() => {
+    agent = createAgent()
   })
 
   afterAll(() => {
