+++ backend/__tests__/acceptance/openapi.spec.js
@@ -13,8 +13,8 @@ const { mockRequest } = require('@gardener-dashboard/request')
 describe('openapi', function () {
   let agent
 
-  beforeAll(async () => {
-    agent = await createAgent()
+  beforeAll(() => {
+    agent = createAgent()
   })
 
   afterAll(() => {
