+++ backend/__tests__/server.spec.js
@@ -9,7 +9,7 @@
 const http = require('http')
 const terminus = require('@godaddy/terminus')
 const metricsApp = require('@gardener-dashboard/monitor')
-let createServer
+const createServer = require('../lib/server')
 
 function createApplication (port, metricsPort) {
   const app = (req, res) => {
@@ -45,9 +45,6 @@ function createApplication (port, metricsPort) {
 }
 
 describe('server', () => {
-  beforeAll(async () => {
-    ({ createServer } = await import('../lib/server.mjs'))
-  })
   const port = 1234
   const metricsPort = 5678
   const mockServer = {
