+++ backend/lib/app.js
@@ -4,20 +4,21 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-import express from 'express'
-import expressStaticGzip from 'express-static-gzip'
-import _ from 'lodash'
-import config from './config/index.js'
-import { resolve, join } from 'path'
-import logger from './logger/index.js'
-import { notFound, renderError, historyFallback, noCache } from './middleware.js'
-import helmet from 'helmet'
-import { hooks, router } from './api.mjs'
-import auth from './auth.js'
-import githubWebhook from './github/webhook/index.js'
-import { healthCheck } from './healthz/index.mjs'
-import { dirname } from 'node:path'
-import { fileURLToPath } from 'node:url'
+'use strict'
+
+const express = require('express')
+const expressStaticGzip = require('express-static-gzip')
+const _ = require('lodash')
+const config = require('./config')
+const { resolve, join } = require('path')
+const logger = require('./logger')
+const { notFound, renderError, historyFallback, noCache } = require('./middleware')
+const helmet = require('helmet')
+const api = require('./api')
+const auth = require('./auth')
+const githubWebhook = require('./github/webhook')
+
+const { healthCheck } = require('./healthz')
 
 const { port, metricsPort } = config
 const periodSeconds = config.readinessProbe?.periodSeconds || 10
@@ -27,8 +28,6 @@ for (const ctor of [Object, Function, Array, String, Number, Boolean]) {
   Object.freeze(ctor)
   Object.freeze(ctor.prototype)
 }
-
-const __dirname = dirname(fileURLToPath(import.meta.url))
 
 // resolve pathnames
 const PUBLIC_DIRNAME = resolve(join(__dirname, '..', 'public'))
@@ -55,7 +54,7 @@ app.set('metricsPort', metricsPort)
 app.set('logger', logger)
 app.set('healthCheck', healthCheck)
 app.set('periodSeconds ', periodSeconds)
-app.set('hooks', hooks)
+app.set('hooks', api.hooks)
 app.set('trust proxy', 1)
 app.set('etag', false)
 app.set('x-powered-by', false)
@@ -69,7 +68,7 @@ if (process.env.NODE_ENV !== 'development') {
 app.use(noCache(STATIC_PATHS))
 app.use('/auth', auth.router)
 app.use('/webhook', githubWebhook.router)
-app.use('/api', router)
+app.use('/api', api.router)
 
 app.use(helmet.xXssProtection())
 app.use(helmet.contentSecurityPolicy({
@@ -104,6 +103,4 @@ app.use(historyFallback(INDEX_FILENAME))
 
 app.use(renderError)
 
-export {
-  app,
-}
+module.exports = app
