+++ backend/package.json
@@ -20,16 +20,16 @@
   "publishConfig": {
     "registry": "https://npm.pkg.github.com/"
   },
-  "main": "server.mjs",
+  "main": "server.js",
   "files": [
     "lib",
-    "server.mjs"
+    "server.js"
   ],
   "scripts": {
-    "serve": "cross-env NODE_ENV=development node server.mjs",
-    "lint": "eslint server.mjs . --max-warnings=0",
+    "serve": "cross-env NODE_ENV=development node server.js",
+    "lint": "eslint server.js . --max-warnings=0",
     "lint-sarif": "yarn lint --format=@microsoft/eslint-formatter-sarif",
-    "test": "cross-env NODE_ENV=test node --experimental-vm-modules $(yarn bin jest)",
+    "test": "cross-env NODE_ENV=test jest",
     "test-coverage": "yarn test --coverage"
   },
   "dependencies": {
@@ -98,8 +98,7 @@
     "verbose": true,
     "coverageDirectory": "coverage",
     "collectCoverageFrom": [
-      "lib/**/*.js",
-      "lib/**/*.mjs"
+      "lib/**/*.js"
     ],
     "testEnvironment": "node",
     "testPathIgnorePatterns": [
